feple-server/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.1
django_celery_results==2.6.0
djangorestframework==3.16.0
idna==3.10
kombu==5.5.3
mysqlclient==2.2.7
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
google-generativeai>=0.4.1,<0.5.0
langchain==0.1.12
langchain-google-genai==0.0.11
sqlalchemy==2.0.28
